+++ crates/store/src/state/writer/apply_block.rs
