+++ crates/store/src/state/writer/apply_proof.rs
